Fix bug preventing moods from changing after setup
Previously when testing with randomized moods, moods were assigned
during setup, and fog was initialized right after, but the fog cam would
never re-render. Now it re-renders at the end of the room scripter. This
does mean the responsibility is on the users to re render after changing
a mood, but changing moods after start does not seem like a use-case at
this point.