Support creating moods on the fly via script
This required changing the moods structure from an array to a list, well
not exactly require, but made it easier.

Now when a mood is created via CreateMood, the Mood asset is created and
it's data is sent to the gpu.

Also now there are two SetMood methods, one that takes an enum for
pre-created moods, and one that just takes a Mood, for one created at
runtime via scripting.