Begin work on custom lights
Custom lights are SO's and get turned into an enum in lighting helper,
they are meant to be added on top of moods in particular positions.

This required an update to the enum code generation so it supports a
generic list of scriptable objects.

Also modelled a quick spot light to use as a first test case.