Error detection and better realtime editing
Associate a dictionary of light position to light types for each room.
This way when moods or lights are edited at runtime, and the lights are
reset, we know which custom lights are associated with that room and we
can re-instance them.

Also added an editor script for the CustomLights so editing them at
playtime also triggers a re-init of light sources.